refactoring in new folders all forms
