+++ moskito-frontend/src/components/Forms/FilterCampsiteForm.js
@@ -3,9 +3,9 @@ import { useState } from 'react'
 import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
 import Feature from './FilterCampsiteCheckbox'
-import Suggestions from './Suggestions'
+import Suggestions from './FormElements/Suggestions'
 import Button from '../Button/Button'
-import InputField from './InputField'
+import InputField from './FormElements/InputField'
 import useForm from '../../hooks/useForm'
 import getInitialCampsiteFilter from '../../services/getInitialCampsiteFilter'
 import { getGeocode, autocomplete} from '../../services/getGeocode'
