+++ moskito-frontend/src/components/Forms/FormElements/Redirection.js
@@ -1,0 +1,17 @@
+import { Redirect } from 'react-router-dom'
+import styled from 'styled-components/macro'
+import Delayed from '../../../hooks/useDelay'
+
+export default function Redirection() {
+    return (
+        <RedParagraph>Registrierung erfolgreich, Du wirst zum Login weitergeleitet.
+            <Delayed waitBeforeShow={2500}>
+                <Redirect to="/login"/>
+            </Delayed>
+        </RedParagraph>
+    ) 
+}
+
+const RedParagraph = styled.p`
+    color: var(--link);
+`
